Minor reorganization of hash refactor
