+++ osquery/CMakeLists.txt
@@ -89,8 +89,6 @@ if(NOT WINDOWS)
   ADD_OSQUERY_LINK_ADDITIONAL("crypto")
   ADD_OSQUERY_LINK_ADDITIONAL("libpthread")
 
-  ADD_OSQUERY_LINK_CORE("crypto")
-
   # Linenoise will be used when compiling osqueryi.
   ADD_OSQUERY_LINK_ADDITIONAL("linenoise")
 endif()
