+++ osquery/config/config.cpp
@@ -22,7 +22,6 @@
 #include <osquery/logger.h>
 #include <osquery/packs.h>
 #include <osquery/registry.h>
-#include <osquery/sha1.h>
 #include <osquery/system.h>
 #include <osquery/tables.h>
 
@@ -716,10 +715,10 @@ void Config::getPerformanceStats(
 
 void Config::hashSource(const std::string& source, const std::string& content) {
   WriteLock wlock(config_hash_mutex_);
-  hash_[source] = getBufferSHA1(&(content.c_str())[0], content.size());
+  hash_[source] = getBufferSHA1(content.c_str(), content.size());
 }
 
-Status Config::getSHA1(std::string& hash) {
+Status Config::genHash(std::string& hash) {
   if (!valid_) {
     return Status(1, "Current config is not valid");
   }
@@ -735,7 +734,7 @@ Status Config::getSHA1(std::string& hash) {
   for (const auto& it : hash_) {
     add(it.second);
   }
-  hash = getBufferSHA1(&buffer[0], buffer.size());
+  hash = getBufferSHA1(buffer.data(), buffer.size());
 
   return Status(0, "OK");
 }
