+++ osquery/tables/forensic/sleuthkit.cpp
@@ -17,11 +17,11 @@
 #include <tsk/libtsk.h>
 
 #include <osquery/filesystem.h>
-#include <osquery/hash.h>
 #include <osquery/logger.h>
 #include <osquery/tables.h>
 
 #include "osquery/core/conversions.h"
+#include "osquery/tables/system/hash.h"
 
 namespace fs = boost::filesystem;
