+++ osquery/tables/system/darwin/keychain_utils.cpp
@@ -14,10 +14,10 @@
 #include <boost/lexical_cast.hpp>
 
 #include <osquery/filesystem.h>
-#include <osquery/hash.h>
 #include <osquery/core.h>
 
 #include "osquery/tables/system/darwin/keychain.h"
+#include "osquery/tables/system/hash.h"
 
 namespace osquery {
 namespace tables {
