+++ osquery/tables/system/posix/smbios_utils.cpp
@@ -8,9 +8,8 @@
  *
  */
 
-#include <osquery/sha1.h>
-
 #include "osquery/tables/system/smbios_utils.h"
+#include "osquery/tables/system/hash.h"
 
 namespace osquery {
 namespace tables {
@@ -124,9 +123,9 @@ void genSMBIOSTable(size_t index,
 
   r["handle"] = BIGINT((unsigned long long)hdr->handle);
   r["header_size"] = INTEGER((unsigned short)hdr->length);
-  r["size"] = INTEGER(size);
-  r["sha1"] = getBufferSHA1(address, size);
 
+  r["size"] = INTEGER(size);
+  r["md5"] = hashFromBuffer(HASH_TYPE_MD5, address, size);
   results.push_back(r);
 }
