+++ specs/posix/acpi_tables.table
@@ -3,7 +3,7 @@ description("Firmware ACPI functional table common metadata and content.")
 schema([
     Column("name", TEXT, "ACPI table name"),
     Column("size", INTEGER, "Size of compiled table data"),
-    Column("sha1", TEXT, "SHA1 hash of table content"),
+    Column("md5", TEXT, "MD5 hash of table content"),
 ])
 implementation("system/acpi_tables@genACPITables")
 fuzz_paths([
